feat(stats subcomponent): popula o subcomponente sales stats que faz parte do componente report
popula o subcomponente sales stats que faz parte do componente report. remaneja e cria novos
arquivos afim de organizar melhor as funcionalidades do componente report
